+++ internal/commands/event/event.go
@@ -18,13 +18,13 @@
 package event
 
 import (
-	"github.com/urfave/cli"
+	"github.com/urfave/cli/v2"
 )
 
-var Command = cli.Command{
+var Command = &cli.Command{
 	Name:  "event",
 	Usage: "Event related sub-command",
-	Subcommands: []cli.Command{
+	Subcommands: []*cli.Command{
 		reportCommand,
 		listCommand,
 	},