+++ pkg/graphql/event/event.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/machinebox/graphql"
 
-	"github.com/urfave/cli"
+	"github.com/urfave/cli/v2"
 
 	api "skywalking.apache.org/repo/goapi/query"
 )