+++ pkg/graphql/metadata/metadata.go
@@ -25,7 +25,7 @@ import (
 	"github.com/apache/skywalking-cli/assets"
 
 	"github.com/machinebox/graphql"
-	"github.com/urfave/cli"
+	"github.com/urfave/cli/v2"
 
 	"github.com/apache/skywalking-cli/pkg/graphql/client"
 )