+++ pkg/graphql/metrics/metrics.go
@@ -19,7 +19,7 @@ package metrics
 
 import (
 	"github.com/machinebox/graphql"
-	"github.com/urfave/cli"
+	"github.com/urfave/cli/v2"
 
 	"github.com/apache/skywalking-cli/assets"
 	"github.com/apache/skywalking-cli/pkg/graphql/client"