+++ pkg/util/template.go
@@ -25,7 +25,7 @@ const TextFormat = `
 {{- $green	:="\x1b[0;32m" -}}	{{- $Green	:="\x1b[1;32m" -}}
 {{- $yellow	:="\x1b[0;33m" -}}	{{- $Yellow	:="\x1b[1;33m" -}}
 {{- $blue	:="\x1b[0;34m" -}}	{{- $Blue	:="\x1b[1;34m" -}}
-{{- $pruple	:="\x1b[0;35m" -}}	{{- $Pruple	:="\x1b[1;35m" -}}
+{{- $purple	:="\x1b[0;35m" -}}	{{- $purple	:="\x1b[1;35m" -}}
 {{- $cyan	:="\x1b[0;36m" -}}	{{- $Cyan	:="\x1b[1;36m" -}}
 {{- $white	:="\x1b[0;37m" -}}	{{- $White	:="\x1b[1;37m" -}}
 {{- $plain	:="\x1b[0m" -}}
@@ -37,8 +37,8 @@ NAME:
 
 USAGE:
 	{{$green}}{{if .UsageText}}{{.UsageText}}{{else}}{{.HelpName}}
-	{{- if .VisibleFlags}}[global options]{{end}}{{if .Commands}} command [command options]{{end}}
-	{{- if .ArgsUsage}}{{.ArgsUsage}}{{else}}[arguments...]{{end}}{{end}}{{$Yellow}}{{if .Version}}{{if not .HideVersion}}
+	{{- if .VisibleFlags}} [global options]{{end}}{{if .Commands}} command [command options]{{end}}
+	{{- if .ArgsUsage}} {{.ArgsUsage}}{{else}} [arguments...]{{end}}{{end}}{{$Yellow}}{{if .Version}}{{if not .HideVersion}}
 
 VERSION:
 	{{$yellow}}{{.Version}}{{end}}{{end}}{{$Yellow}}{{if .Description}}
@@ -66,11 +66,11 @@ COPYRIGHT:
 
 const CommandHelpTemplate = TextFormat + `{{$Green -}}
 NAME:
-	{{$green}}{{.HelpName}} - {{.Usage}}{{$Green}}
+	{{$green}}{{.HelpName}}{{if .VisibleFlags}} [command options]{{end}}{{- if .ArgsUsage}} {{.ArgsUsage}}{{else}}{{end}} - {{.Usage}}{{$Green}}
 
 USAGE:
-	{{$green}}{{if .UsageText}}{{.UsageText}}{{else}}{{.HelpName}}{{if .VisibleFlags}} [command options]{{end}} 
-	{{- if .ArgsUsage}}{{.ArgsUsage}}{{else}}[arguments...]{{end}}{{end}}{{$Green}}{{if .Category}}
+	{{$green}}{{if .UsageText}}{{.UsageText | nindent 2 | trim}}{{else}}{{.HelpName}}{{if .VisibleFlags}} [command options]{{end}} 
+	{{- if .ArgsUsage}} {{.ArgsUsage}}{{end}}{{end}}{{$Green}}{{if .Category}}
 
 CATEGORY:
 	{{$green}}{{.Category}}{{end}}{{$Blue}}{{if .Description}}
@@ -85,19 +85,19 @@ OPTIONS:
 
 const SubcommandHelpTemplate = TextFormat + `{{$Green -}}
 NAME:
-   {{$green}}{{.HelpName}} - {{if .Description}}{{.Description}}{{else}}{{.Usage}}{{end}}{{$Green}}
+	{{$green}}{{.HelpName}} - {{if .Description}}{{.Description}}{{else}}{{.Usage}}{{end}}{{$Green}}
 
 USAGE:
-   {{$green}}{{if .UsageText}}{{.UsageText}}{{else}}{{.HelpName}} command{{if .VisibleFlags}} [command options]{{end}}
-   {{- if .ArgsUsage}}{{.ArgsUsage}}{{else}}[arguments...]{{end}}{{end}}{{$Blue}}
+	{{$green}}{{if .UsageText}}{{.UsageText | nindent 2 | trim}}{{else}}{{.HelpName}} command{{if .VisibleFlags}} [command options]{{end}}
+	{{- if .ArgsUsage}} {{.ArgsUsage}}{{else}} [arguments...]{{end}}{{end}}{{$Blue}}
 
 COMMANDS:{{range .VisibleCategories}}{{if .Name}}
 
-   {{.Name}}:{{range .VisibleCommands}}
-   {{$Blue}}{{join .Names ", "}}{{$blue}}{{"\t"}}{{.Usage}}{{end}}{{else}}{{range .VisibleCommands}}
-   {{$Blue}}{{join .Names ", "}}{{$blue}}{{"\t"}}{{.Usage}}{{end}}{{end}}{{end}}{{if .VisibleFlags}}{{$Cyan}}
+	{{.Name}}:{{range .VisibleCommands}}
+	{{$Blue}}{{join .Names ", "}}{{$blue}}{{"\t"}}{{.Usage}}{{end}}{{else}}{{range .VisibleCommands}}
+	{{$Blue}}{{join .Names ", "}}{{$blue}}{{"\t"}}{{.Usage}}{{end}}{{end}}{{end}}{{if .VisibleFlags}}{{$Cyan}}
 
 OPTIONS:
-   {{$cyan}}{{range .VisibleFlags}}{{.}}
-   {{end}}{{end}}{{$plain}}
+	{{$cyan}}{{range .VisibleFlags}}{{.}}
+	{{end}}{{end}}{{$plain}}
 `